Support splunk-[heavy]forwarder pulls by digest
Add two fields to the SplunkForwarder CRD: `imageDigest` and
`heavyForwarderDigest`. If `imageDigest` is specified, the DaemonSet's
container image pull spec is constructed as `{image}@{imageDigest}`.
Similarly, if `heavyForwarderDigest` is specified, the Deployment's
container image pull spec is constructed as
`{heavyForwarderImage}@{heavyForwarderDigest}`. In either case, if
`imageTag` is present it is ignored.

OSD-7065