s-(h)f images are now built externally
Since https://github.com/openshift/splunk-forwarder-images/pull/1,
splunk-forwarder and splunk-heavyforwarder images are built by the
splunk-forwarder-images repository. This commit removes all the logic
that was building them here.

It also introduces a convenience script and `make` target to discover
and populate the digests for those images in the OLM template.

OSD-7272