add init container for secret handling
The secret needs to be owned by `splunkfwd:splunkfwd` on the filesystem
because for some reason known only to Splunk, the Universal Forwarder
overwrites the original outputs.conf file. Mounting the secret as a file
directly makes it owned by the root user, so overwriting is not
possible. The init container copies the secret to an ephemeral volume
and changes the owner in order to deal with that.