remove invalid test
The testing function calls `Reconcile()` directly, which bypasses the
predicate that tests the name and makes sure reconciliation doesn't run
for secrets we don't care about.

This test was passing before because it would get to the logic that
checked if the secret existed and stopped reconciliation if it didn't,
which did not return an error.