sync HEC token to all clusters
Since HEC token authentication takes priority over mTLS auth, this will
mean the Splunk forwarder will switch over to HTTP mode for all OSD and
ROSA Classic clusters. Until the work to create an operator to manage
the token lifecycle is completed the clusters will be sharing a single
HEC token for authentication, but that is the same state as the mTLS
certificate so there is no added risk.

Signed-off-by: Jason Healy <jahealy@redhat.com>